// repos/konraddb/ui_private/src/definitions/types/backdrops.ts
export enum Backdrops {
  LOGIN = "LOGIN",
  CREATE_NEW_API_KEY = "CREATE_NEW_API_KEY",
  DELETE_API_KEY = "DELETE_API_KEY",
  CONTACT_US = "CONTACT_US",

  REQUEST_WITHDRAWAL = "REQUEST_WITHDRAWAL",

  REQUEST_WHITELIST_ADDRESS = "REQUEST_WHITELIST_ADDRESS",

  CANCEL_ORDERS = "CANCEL_ORDERS",

  CHANGE_ROLE_BACKDROP = "CHANGE_ROLE_BACKDROP",
  INVITE_USER_BACKDROP = "INVITE_USER_BACKDROP",
  DELETE_ADDRESS = "DELETE_ADDRESS",
  ADDRESS_LABEL_CHANGE = "ADDRESS_LABEL_CHANGE",
  REQUEST_POOL_DEPOSIT = "REQUEST_POOL_DEPOSIT",
}